refactor_for_testability
