+++ tests/modeling/tpc_ds_duckdb/aggregates/zquery_timing_amd64-intel64_family_6_model_183_stepping_1_genuineintel-16.log
@@ -1,4 +1,4 @@
 [query_03]
-parse_time = 0.033159
-exec_time = 0.029805
-comp_time = 0.101366
+parse_time = 0.030927
+exec_time = 0.025167
+comp_time = 0.092646
