+++ tests/modeling/tpc_ds_duckdb/zquery_timing_amd64-intel64_family_6_model_183_stepping_1_genuineintel-16.log
@@ -1,124 +1,124 @@
 [query_01]
-parse_time = 0.083158
-exec_time = 0.08899
-comp_time = 0.03186
+parse_time = 0.094887
+exec_time = 0.087546
+comp_time = 0.031122
 
 [query_02]
-parse_time = 0.286068
-exec_time = 0.086957
-comp_time = 0.094979
+parse_time = 0.309827
+exec_time = 0.122748
+comp_time = 0.09189
 
 [query_03]
-parse_time = 0.037796
-exec_time = 0.032438
-comp_time = 0.036914
+parse_time = 0.03948
+exec_time = 0.035082
+comp_time = 0.031706
 
 [query_06]
-parse_time = 0.062282
-exec_time = 0.033797
-comp_time = 0.047203
+parse_time = 0.080871
+exec_time = 0.032373
+comp_time = 0.049234
 
 [query_07]
-parse_time = 0.046723
-exec_time = 0.114847
-comp_time = 0.084821
+parse_time = 0.049159
+exec_time = 0.114406
+comp_time = 0.092751
 
 [query_08]
-parse_time = 0.0998
-exec_time = 0.050864
-comp_time = 0.122065
+parse_time = 0.10074
+exec_time = 0.05044
+comp_time = 0.121982
 
 [query_10]
-parse_time = 0.228907
-exec_time = 0.180674
-comp_time = 0.242532
+parse_time = 0.224238
+exec_time = 0.180066
+comp_time = 0.26179
 
 [query_12]
-parse_time = 0.063481
-exec_time = 0.033477
-comp_time = 0.106414
+parse_time = 0.077162
+exec_time = 0.04183
+comp_time = 0.126043
 
 [query_15]
-parse_time = 0.03933
-exec_time = 0.020149
-comp_time = 0.140963
+parse_time = 0.053979
+exec_time = 0.018973
+comp_time = 0.140642
 
 [query_16]
-parse_time = 0.124089
-exec_time = 0.146923
-comp_time = 0.020214
+parse_time = 0.124258
+exec_time = 0.141383
+comp_time = 0.019549
 
 [query_20]
-parse_time = 0.057977
-exec_time = 0.076764
-comp_time = 0.132803
+parse_time = 0.055344
+exec_time = 0.066731
+comp_time = 0.117438
 
 [query_21]
-parse_time = 0.01665
-exec_time = 0.023557
-comp_time = 0.024928
+parse_time = 0.018031
+exec_time = 0.023706
+comp_time = 0.023666
 
 [query_22]
-parse_time = 0.42315
-exec_time = 0.383677
-comp_time = 0.802893
+parse_time = 1.269483
+exec_time = 0.401332
+comp_time = 0.814309
 
 [query_24]
-parse_time = 0.104903
-exec_time = 0.193135
-comp_time = 0.174558
+parse_time = 0.11442
+exec_time = 0.230021
+comp_time = 0.183518
 
 [query_25]
-parse_time = 0.203972
-exec_time = 0.11937
-comp_time = 0.045725
+parse_time = 0.245857
+exec_time = 0.146282
+comp_time = 0.04913
 
 [query_26]
-parse_time = 0.04282
-exec_time = 0.07787
-comp_time = 0.061977
+parse_time = 0.047781
+exec_time = 0.079631
+comp_time = 0.06217
 
 [query_30]
-parse_time = 0.114736
-exec_time = 0.122662
-comp_time = 0.048189
+parse_time = 0.125266
+exec_time = 0.13588
+comp_time = 0.045789
 
 [query_32]
-parse_time = 0.060459
-exec_time = 0.010075
-comp_time = 0.007245
+parse_time = 0.062862
+exec_time = 0.011492
+comp_time = 0.006314
 
 [query_42]
-parse_time = 0.595779
-exec_time = 0.025088
-comp_time = 0.024759
+parse_time = 0.049486
+exec_time = 0.027106
+comp_time = 0.02704
 
 [query_55]
-parse_time = 0.037965
-exec_time = 0.028706
-comp_time = 0.025501
+parse_time = 0.039602
+exec_time = 0.026381
+comp_time = 0.024532
 
 [query_95]
-parse_time = 0.094288
-exec_time = 0.126694
-comp_time = 0.517112
+parse_time = 0.953062
+exec_time = 0.140251
+comp_time = 0.562263
 
 [query_97]
-parse_time = 0.156502
-exec_time = 0.223412
-comp_time = 0.090547
+parse_time = 0.217524
+exec_time = 0.242053
+comp_time = 0.103125
 
 ["query_97.1"]
-parse_time = 0.109867
-exec_time = 1.012204
-comp_time = 0.116316
+parse_time = 0.128833
+exec_time = 1.162842
+comp_time = 0.11708
 
 [query_98]
-parse_time = 0.088004
-exec_time = 0.047235
-comp_time = 0.210617
+parse_time = 0.124205
+exec_time = 0.059557
+comp_time = 0.240875
 
 [query_99]
-parse_time = 0.070156
-exec_time = 0.099379
-comp_time = 0.073402
+parse_time = 0.07627
+exec_time = 0.101067
+comp_time = 0.069631
