+++ tests/scripts/config_directory/raw/users_with_pk.preql
@@ -1,5 +1,5 @@
 # Datasource ingested from users_with_pk
-# Generated on 2026-04-27 14:18:52.087933
+# Generated on 2026-04-27 15:25:08.373163
 import std.net;
 
 key user_id int;
